+++ tests/test_basic_example.py
@@ -42,4 +42,4 @@ def test_basic_example(testdir):
     utils.add_version_constraint_to_project(testdir.tmpdir)
 
     result = testdir.runpytest("tests/test_quickstart.py")
-    result.assert_outcomes(passed=6)
+    result.assert_outcomes(passed=8)