+++ tests/test_containerized_orchestrator.py
@@ -89,4 +89,4 @@ def test_basic_example(testdir: Testdir):
     utils.add_version_constraint_to_project(testdir.tmpdir)
 
     result = testdir.runpytest("tests/test_quickstart.py", "--lsm-ctr")
-    result.assert_outcomes(passed=6)
+    result.assert_outcomes(passed=8)